+++ bluesky/src/commonMain/kotlin/work/socialhub/planetlink/bluesky/model/BlueskyComment.kt
@@ -3,7 +3,6 @@ package work.socialhub.planetlink.bluesky.model
 import work.socialhub.kbsky.model.atproto.repo.RepoStrongRef
 import work.socialhub.kbsky.util.ATUriParser
 import work.socialhub.planetlink.micro.MicroBlogComment
-import work.socialhub.planetlink.model.Comment
 import work.socialhub.planetlink.model.Identify
 import work.socialhub.planetlink.model.Reaction
 import work.socialhub.planetlink.model.Service
